+++ ci/test_rasters.py
@@ -32,8 +32,6 @@ from sertit.rasters import (
     INT8_NODATA,
     UINT8_NODATA,
     UINT16_NODATA,
-    any_raster_to_xr_ds,
-    get_nodata_value_from_dtype,
     get_nodata_value_from_xr,
 )
 from sertit.vectors import EPSG_4326
@@ -696,11 +694,6 @@ def test_write(dtype, nodata_val, tmp_path, xda):
         )
         _test_raster_after_write(test_path, dtype, nodata_val)
 
-    # test deprecation warning
-    test_deprecated_path = os.path.join(tmp_path, "test_depr.tif")
-    with pytest.deprecated_call():
-        rasters.write(xda, path=test_deprecated_path, dtype=dtype)
-
 
 def test_dim():
     """Test on BEAM-DIMAP function"""
@@ -919,51 +912,6 @@ def test_rasterize(tmp_path, raster_path):
     rasters.write(rast, out_path, dtype=np.uint8, nodata=255)
 
     ci.assert_raster_almost_equal(raster_true_path, out_path, decimal=4)
-
-
-@s3_env
-def test_decorator_deprecation(raster_path):
-    from sertit.rasters import path_xarr_dst
-
-    @any_raster_to_xr_ds
-    def _ok_rasters(xds):
-        assert isinstance(xds, xr.DataArray)
-        return xds
-
-    @path_xarr_dst
-    def _depr_rasters(xds):
-        assert isinstance(xds, xr.DataArray)
-        return xds
-
-    # Not able to warn deprecation from inside the decorator
-    xr.testing.assert_equal(_ok_rasters(raster_path), _depr_rasters(raster_path))
-
-
-def test_get_nodata_deprecation():
-    """Test deprecation of get_nodata_value"""
-    # Test deprecation
-    for dtype in [
-        np.uint8,
-        np.int8,
-        np.uint16,
-        np.uint32,
-        np.int32,
-        np.int64,
-        np.uint64,
-        int,
-        "int",
-        np.int16,
-        np.float32,
-        np.float64,
-        float,
-        "float",
-    ]:
-        with pytest.deprecated_call():
-            from sertit.rasters import get_nodata_value
-
-            ci.assert_val(
-                get_nodata_value_from_dtype(dtype), get_nodata_value(dtype), dtype
-            )
 
 
 @s3_env